Add files via upload
Working with the robot involved programming the robotic manipulator and integrating it with external devices such as sensors, controllers, solenoid valves and a Siemens S7-1200 PLC. The exact interaction between these components and the robot was programmed in TIA Portal, which enabled the pucks to be picked correctly and segregated based on appearance, thanks to the sensors and laser. The pucks were segregated into two ramps, where they were counted. One contained those with holes from the center or side, and the other contained those without.